remove public API `ConditionEvents.handleViolations(..)`
To make `ConditionEvents` easier to maintain we should slim down the interface as much as possible. Users should rely on `EvaluationResult` to handle specific violations, not on `ConditionEvents`.
