+++ archunit/src/test/java/com/tngtech/archunit/lang/EvaluationResultTest.java
@@ -1,12 +1,15 @@
 package com.tngtech.archunit.lang;
 
+import java.util.ArrayList;
 import java.util.Collection;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 
+import com.google.common.base.Joiner;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableSet;
+import com.google.common.collect.Iterables;
 import com.tngtech.archunit.base.HasDescription;
 import org.junit.Test;
 
@@ -78,6 +81,78 @@ public class EvaluationResultTest {
         assertThat(filtered.getFailureReport().getDetails()).containsOnly("keep first line1", "keep second line1", "keep second line2");
     }
 
+    @Test
+    @SuppressWarnings("Convert2Lambda") // We need the reified type here, so we cannot convert it to a lambda
+    public void handleViolations_reports_only_violations_referring_to_the_correct_type() {
+        EvaluationResult result = evaluationResultWith(
+                SimpleConditionEvent.satisfied(new CorrectType("do not handle"), "I'm not violated"),
+                SimpleConditionEvent.violated(new WrongType(), "I'm violated, but wrong type"),
+                SimpleConditionEvent.violated(new WrongSupertype(), "I'm violated, but wrong type"),
+                SimpleConditionEvent.violated(new CorrectType("handle type"), "I'm violated and correct type"),
+                SimpleConditionEvent.violated(new CorrectSubtype("handle sub type"), "I'm violated and correct sub type"));
+
+        final Set<String> handledFailures = new HashSet<>();
+        result.handleViolations(new ViolationHandler<CorrectType>() {
+            @Override
+            public void handle(Collection<CorrectType> violatingObjects, String message) {
+                handledFailures.add(Joiner.on(", ").join(violatingObjects) + ": " + message);
+            }
+        });
+
+        assertThat(handledFailures).containsOnly(
+                "handle type: I'm violated and correct type",
+                "handle sub type: I'm violated and correct sub type");
+    }
+
+    @Test
+    public void handles_erased_generics_as_upper_bound() {
+        EvaluationResult result = evaluationResultWith(
+                SimpleConditionEvent.violated(new CorrectType("ignore"), "correct"),
+                SimpleConditionEvent.violated(new WrongType(), "wrong"));
+
+        Set<String> handledFailureMessages = new HashSet<>();
+        result.handleViolations(genericBoundByCorrectType(handledFailureMessages));
+
+        assertThat(handledFailureMessages).containsOnly("correct");
+
+        handledFailureMessages = new HashSet<>();
+        result.handleViolations(unboundGeneric(handledFailureMessages));
+
+        assertThat(handledFailureMessages).containsOnly("correct", "wrong");
+    }
+
+    @SuppressWarnings("Convert2Lambda") // to retrieve the type information ViolationHandler may not be converted to a Lambda
+    private <T extends CorrectType> ViolationHandler<?> genericBoundByCorrectType(final Set<String> handledFailureMessages) {
+        return new ViolationHandler<T>() {
+            @Override
+            public void handle(Collection<T> violatingObjects, String message) {
+                handledFailureMessages.add(message);
+            }
+        };
+    }
+
+    @SuppressWarnings("Convert2Lambda") // to retrieve the type information ViolationHandler may not be converted to a Lambda
+    private <T> ViolationHandler<?> unboundGeneric(final Set<String> handledFailureMessages) {
+        return new ViolationHandler<T>() {
+            @Override
+            public void handle(Collection<T> violatingObjects, String message) {
+                handledFailureMessages.add(message);
+            }
+        };
+    }
+
+    @Test
+    public void can_handle_with_generic_superclasses() {
+        EvaluationResult result = evaluationResultWith(
+                SimpleConditionEvent.violated(new Object(), "ignore"),
+                SimpleConditionEvent.violated("correct", "ignore"));
+
+        StringHandler handler = new StringHandler();
+        result.handleViolations(handler);
+
+        assertThat(handler.getRecorded()).containsOnly("correct");
+    }
+
     private EvaluationResult evaluationResultWith(ConditionEvent... events) {
         return new EvaluationResult(hasDescription("unimportant"), events(events), MEDIUM);
     }
@@ -99,6 +174,48 @@ public class EvaluationResultTest {
 
     private HasDescription hasDescription(final String description) {
         return () -> description;
+    }
+
+
+    private static class BaseHandler<T> implements ViolationHandler<T> {
+        private final List<T> recorded = new ArrayList<>();
+
+        @Override
+        public void handle(Collection<T> violatingObjects, String message) {
+            recorded.add(Iterables.getOnlyElement(violatingObjects));
+        }
+
+        List<T> getRecorded() {
+            return recorded;
+        }
+    }
+
+    private static class StringHandler extends BaseHandler<String> {
+    }
+
+    private static class CorrectSubtype extends CorrectType {
+        CorrectSubtype(String message) {
+            super(message);
+        }
+    }
+
+    static class CorrectType extends WrongSupertype {
+        String message;
+
+        CorrectType(String message) {
+            this.message = message;
+        }
+
+        @Override
+        public String toString() {
+            return message;
+        }
+    }
+
+    private static class WrongType {
+    }
+
+    private static class WrongSupertype {
     }
 
     private static class TestEvent implements ConditionEvent {
